completed without hilt

+++ app/src/main/java/com/example/yoospace_android/utils/generateGradient.kt
@@ -3,7 +3,6 @@ package com.example.yoospace_android.utils
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import kotlin.math.absoluteValue
 
 fun generateGradient(userName: String): Brush {
     fun hashCode(str: String): Int {
@@ -12,7 +11,7 @@ fun generateGradient(userName: String): Brush {
 
     val hash = hashCode(userName)
     val hue1 = (hash % 360).toFloat()
-    val hue2 = ((hash * 1.5f) % 360).toFloat()
+    val hue2 = ((hash * 1.5f) % 360)
 
     val color1 = Color.hsl(hue1, 0.7f, 0.6f)
     val color2 = Color.hsl(hue2, 0.7f, 0.4f)
